refactor(db): move manifest implementation details into common/manifest/
common/ now keeps only the public API (manifest_codec.h, manifest_enum.h). The codec implementation, the pb_wire reader/writer and a new README documenting the wire format and its compatibility rules (how to add fields, why retired field numbers must never be reused) live in common/manifest/.

Also split the long Reader::Next()/ReadVarintRaw() implementations out of pb_wire.h into pb_wire.cc, keeping only declarations and short inline methods in the header.
