+++ src/db/CMakeLists.txt
@@ -1,10 +1,6 @@
 include(${PROJECT_ROOT_DIR}/cmake/bazel.cmake)
 include(${PROJECT_ROOT_DIR}/cmake/option.cmake)
 
-# NOTE: the manifest on-disk format is read and written by
-# src/db/index/common/manifest_codec.{h,cc}; there is no protobuf/protoc
-# dependency any more.
-
 cc_directory(common)
 cc_directory(index)
 cc_directory(sqlengine)
